requirements.txt
streamlit==1.45.0
pandas==2.2.3
pdfplumber==0.11.6
openai==1.78.0
google-genai==1.15.0
openpyxl==3.1.5